ref: tidy up utils
